Improved faction check
F2 now detects if a unit's faction is different to the faction of it's
group leader and if so treats the unit as if it was belonging to the
FTL's faction.
UN Medics are now CDF instead of PMC.

This affects briefing, groupMarkers and gear assignment.

As Engineers are their own groups this still doesn't fix the Engineers
for Russia, TK Local and UN. Placing a dummy unit as groupleader with
probabilitiy 0% and a "don't delete me" description would solve that
issue.

Note:
This will "break" missions with groups of mixed factions and where
varying briefings/gear/markers are desired. It's easily fixed by
commenting the lines in the varying scripts and adapting the clauses.